refactor: 데스크탑 헤더컴포넌트 rsc와 클라이언트 컴포넌트 분리하여 성능개선
