+++ src/components/common/mobile-header.tsx
